requirements.txt
transformers==4.29.2
gradio==2.2.14
torch==1.11.0
Pillow==9.0.1
timm==0.9.0